requirements.txt

streamlit
ultralytics
streamlit-webrtc
av
numpy
Pillow
opencv-python-headless
